Re-enable all workflows and neaten code.
